Refactor Dashboard and add comments
